+++ crates/wdl-analysis/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Added
+
+* Added an `ArrayType::unqualified` method to cheaply drop the `+` qualifier ([#529](https://github.com/stjude-rust-labs/wdl/pull/529)).
+
 #### Changed
 
 * The `UnusedCall` rule no longer emits a diagnostic for tasks/workflows called if they have an empty or missing `output` section ([#532](https://github.com/stjude-rust-labs/wdl/pull/532)).